Update Object Velocities Using Gravity
A couple changes:
(1) No longer track direction and speed of an object separately, just roll it into a velocity vector.  This makes calculations simpler.
(2) Collision.vector contains the impulse vector (ie: the change in velocity due to the collision) instead of the direction vector.
(3) Store a deltaT on game, which is the time elapsed since last render.
(4) Add gravity constant to World, which is used to calculate the change in velocity due to gravity.
(5) In update(), add the velocity change vector to the current velocity vector.

Sorry for the whitespace changes.  Emacs removed trailing whitespace and I didn't notice until it was too annoying to fix.